refactor: errors into proper exports
